third_party/WebKit/LayoutTests/fast/dom/dom-attribute-on-prototype-chain-expected.txt

PASS div.hasOwnProperty("id") is false
PASS div.__proto__.__proto__.__proto__.hasOwnProperty("id") is true
PASS desc.get instanceof Function is true
PASS desc.set instanceof Function is true
PASS desc.enumerable is true
PASS desc.configurable is true
PASS xhr = new XMLHttpRequest(); xhr.__proto__ = HTMLDivElement.prototype; xhr.id threw exception TypeError: Illegal invocation.
PASS obj = Object.create(div); obj.id threw exception TypeError: Illegal invocation.
PASS desc.get.call({}) threw exception TypeError: Illegal invocation.
PASS desc.set.call(a, "abc") is undefined.
PASS desc.get.call(a) is "abc"
PASS successfullyParsed is true

TEST COMPLETE
